refactor(PVK-48): Split traits and implementations for nas service
