+++ backend/homelab_back/apps/nas-server/src/service/contract/global_file_service.rs
@@ -1,0 +1,8 @@
+use async_trait::async_trait;
+use homelab_core::global_file::GlobalFile;
+use crate::helpers::data_error::DataError;
+
+#[async_trait]
+pub trait GlobalFileService: Send + Sync {
+    async fn get_all(&self) -> Result<Vec<GlobalFile>, DataError>;
+}
