+++ backend/homelab_back/apps/nas-server/src/service/contract/mod.rs
@@ -1,0 +1,11 @@
+pub mod clean_up_service;
+pub mod file_label_service;
+pub mod file_read_service;
+pub mod file_write_service;
+pub mod folder_read_service;
+pub mod folder_write_service;
+pub mod global_file_service;
+pub mod label_service;
+pub mod preview_service;
+pub mod shared_file_service;
+pub mod sp_service;
