+++ backend/homelab_back/apps/nas-server/src/service/contract/preview_service.rs
@@ -1,0 +1,6 @@
+use std::path::PathBuf;
+use homelab_core::file::File;
+
+pub trait PreviewService {
+    fn spawn_generation(file: File, storage_path: PathBuf);
+}
