--- backend/homelab_back/apps/nas-server/src/service/folder_service.rs
@@ -1,227 +1,0 @@
-use crate::data::create_folder_command::CreateFolderCommand;
-use crate::data::move_folder_command::MoveFolderCommand;
-use crate::data::update_folder_name_command::UpdateFolderNameCommand;
-use crate::db::folder_repository::FolderRepository;
-use crate::helpers::data_error::DataError;
-use async_recursion::async_recursion;
-use async_trait::async_trait;
-use homelab_core::file::{File, FileType};
-use homelab_core::folder::Folder;
-use std::sync::Arc;
-use derive_new::new;
-use uuid::Uuid;
-use homelab_core::events::{DeletionType, TrashCleanUpTriggeredEvent};
-use crate::events::rabbitmq::RabbitMqPublisher;
-
-#[async_trait]
-pub trait FolderService: Send + Sync {
-    async fn get_root(&self, user_id: Uuid) -> Result<Option<Folder>, DataError>;
-    async fn get_by_id(&self, folder_id: Uuid) -> Result<Option<Folder>, DataError>;
-    async fn get_children_by_id(&self, folder_id: Uuid) -> Result<Vec<Folder>, DataError>;
-    async fn search_folder(&self, search_query: String) -> Result<Vec<Folder>, DataError>;
-    async fn filter_files_by_folder(
-        &self,
-        file_types: &[FileType],
-        folder_id: Uuid,
-    ) -> Result<Vec<File>, DataError>;
-    async fn get_folder_path(&self, folder_id: Uuid) -> Result<String, DataError>;
-    async fn get_by_folder(&self, folder_id: Uuid) -> Result<Vec<File>, DataError>;
-    async fn update_folder_name(
-        &self,
-        command: UpdateFolderNameCommand,
-        folder_id: Uuid,
-    ) -> Result<Folder, DataError>;
-    async fn trash_chosen_folders(&self, folder_ids: &[Uuid]) -> Result<(), DataError>;
-    async fn trash(&self, folder_id: Uuid) -> Result<(), DataError>;
-    async fn create(&self, command: CreateFolderCommand) -> Result<Folder, DataError>;
-    async fn move_folder(&self, command: MoveFolderCommand) -> Result<Folder, DataError>;
-    async fn get_trash_files(&self, folder_id: Uuid) -> Result<Vec<File>, DataError>;
-    async fn get_trash_subfolder(&self, folder_id: Uuid) -> Result<Vec<Folder>, DataError>;
-    async fn get_deleted_folders(&self, user_id: Uuid) -> Result<Vec<Folder>, DataError>;
-    async fn clean_up_trash(&self, user_id: Uuid) -> Result<(), DataError>;
-    async fn permanently_delete_folder(&self, folder_id: Uuid, user_id: Uuid) -> Result<(), DataError>;
-    async fn restore_deleted_folder(&self, folder_id: Uuid) -> Result<(), DataError>;
-}
-
-#[derive(new)]
-pub struct FolderServiceImpl {
-    folder_repo: Arc<dyn FolderRepository>,
-    publisher: Arc<RabbitMqPublisher>,
-}
-
-impl FolderServiceImpl {
-    #[async_recursion]
-    async fn get_parent_folder_name(&self, f_id: Uuid) -> Result<String, DataError> {
-        let f = self
-            .folder_repo
-            .get_by_id(f_id)
-            .await?
-            .ok_or_else(|| DataError::EntityNotFoundException("Folder".to_string()))?;
-
-        if let Some(parent_id) = f.parent_folder_id {
-            let parent_path = self.get_parent_folder_name(parent_id).await?;
-            Ok(format!("{}/{}", parent_path, f.name))
-        } else {
-            Ok(f.name)
-        }
-    }
-}
-
-#[async_trait]
-impl FolderService for FolderServiceImpl {
-    async fn get_root(&self, user_id: Uuid) -> Result<Option<Folder>, DataError> {
-        self.folder_repo.get_root(user_id).await
-    }
-
-    async fn get_by_id(&self, folder_id: Uuid) -> Result<Option<Folder>, DataError> {
-        self.folder_repo.get_by_id(folder_id).await
-    }
-
-    async fn get_children_by_id(&self, folder_id: Uuid) -> Result<Vec<Folder>, DataError> {
-        self.folder_repo.get_children_by_id(folder_id).await
-    }
-
-    async fn search_folder(&self, search_query: String) -> Result<Vec<Folder>, DataError> {
-        self.folder_repo
-            .search_by_name(format!("%{}%", search_query))
-            .await
-    }
-
-    async fn filter_files_by_folder(
-        &self,
-        file_types: &[FileType],
-        folder_id: Uuid,
-    ) -> Result<Vec<File>, DataError> {
-        self.folder_repo
-            .filter_files_in_folder(file_types, folder_id)
-            .await
-    }
-
-    async fn get_folder_path(&self, folder_id: Uuid) -> Result<String, DataError> {
-        let path = self.get_parent_folder_name(folder_id).await?;
-        Ok(path)
-    }
-
-    async fn get_by_folder(&self, folder_id: Uuid) -> Result<Vec<File>, DataError> {
-        self.folder_repo.get_by_folder_id(folder_id).await
-    }
-
-    async fn update_folder_name(
-        &self,
-        command: UpdateFolderNameCommand,
-        folder_id: Uuid,
-    ) -> Result<Folder, DataError> {
-        let mut folder: Folder = self
-            .folder_repo
-            .get_by_id(folder_id)
-            .await?
-            .ok_or_else(|| DataError::EntityNotFoundException("File".to_string()))?;
-
-        folder.rename(command.new_name);
-
-        self.folder_repo.update_folder(folder).await
-    }
-
-    async fn trash_chosen_folders(&self, folder_ids: &[Uuid]) -> Result<(), DataError> {
-        self.folder_repo.mark_folders_deleted(folder_ids).await
-    }
-
-    async fn trash(&self, folder_id: Uuid) -> Result<(), DataError> {
-        self.folder_repo.mark_folder_deleted(folder_id).await
-    }
-
-    async fn create(&self, command: CreateFolderCommand) -> Result<Folder, DataError> {
-        let f = Folder::new(
-            Uuid::new_v4(),
-            Some(command.parent_folder_id),
-            command.name,
-            command.owner_id,
-        );
-
-        self.folder_repo.create(f).await
-    }
-
-    async fn move_folder(&self, command: MoveFolderCommand) -> Result<Folder, DataError> {
-        let mut folder = self
-            .folder_repo
-            .get_by_id(command.folder_id)
-            .await?
-            .ok_or_else(|| DataError::EntityNotFoundException("Folder".to_string()))?;
-
-        folder.update_parent_folder(command.target_folder);
-
-        Ok(self.folder_repo.update_folder(folder).await?)
-    }
-
-    async fn get_trash_files(&self, folder_id: Uuid) -> Result<Vec<File>, DataError> {
-        self.folder_repo.get_trash_file_for_folder(folder_id).await
-    }
-
-    async fn get_trash_subfolder(&self, folder_id: Uuid) -> Result<Vec<Folder>, DataError> {
-        self.folder_repo.get_trash_subfolder_for_folder(folder_id).await
-    }
-
-    async fn get_deleted_folders(&self, user_id: Uuid) -> Result<Vec<Folder>, DataError> {
-        self.folder_repo.get_deleted_folders(user_id).await
-    }
-
-    async fn clean_up_trash(&self, user_id: Uuid) -> Result<(), DataError> {
-        let event = TrashCleanUpTriggeredEvent::new(
-            user_id,
-            DeletionType::File,
-            None
-        );
-        
-        self
-            .publisher
-            .publish(&event)
-            .await
-            .map_err(|e| DataError::MessageQueueError(format!("TrashCleanUpTriggeredEvent to clean up user's trash {:?}", e)))?;
-        
-        Ok(())
-    }
-
-    async fn permanently_delete_folder(&self, folder_id: Uuid, user_id: Uuid) -> Result<(), DataError> {
-        println!("Cleaning up deleted folder: {:?}", folder_id);
-        let folder = self.folder_repo.get_by_id(folder_id).await?
-            .ok_or_else(|| DataError::EntityNotFoundException("File".to_string()))?;
-        
-        if folder.owner_id != user_id {
-            return Err(DataError::InvalidDataError("Cannot delete folder that does not belong to this user".to_string()));
-        }
-        
-        let event = TrashCleanUpTriggeredEvent::new(
-            user_id,
-            DeletionType::Folder,
-            Some(folder_id)
-        );
-
-        self
-            .publisher
-            .publish(&event)
-            .await
-            .map_err(|e| DataError::MessageQueueError(format!("TrashCleanUpTriggeredEvent to delete a folder {:?}", e)))?;
-
-        Ok(())
-    }
-
-    async fn restore_deleted_folder(&self, folder_id: Uuid) -> Result<(), DataError> {
-        let mut folder = self.folder_repo.get_by_id(folder_id).await?
-            .ok_or_else(|| DataError::EntityNotFoundException("Folder".to_string()))?;
-
-        let parent_folder = self.folder_repo.get_by_id(folder.parent_folder_id.unwrap())
-            .await?
-            .ok_or_else(|| DataError::EntityNotFoundException("Folder".to_string()))?;
-
-        let root_folder = self.folder_repo.get_root(folder.owner_id)
-            .await?.ok_or_else(|| DataError::EntityNotFoundException("Folder".to_string()))?;
-
-        if parent_folder.is_deleted {
-            folder.update_parent_folder(root_folder.id);
-
-            let _f = self.folder_repo.update_folder(folder).await?;
-        }
-
-        self.folder_repo.restore_deleted_folder(folder_id).await
-    }
-}
