+++ backend/homelab_back/apps/nas-server/src/service/impl/sp_service_impl.rs
@@ -7,14 +7,7 @@ use homelab_core::storage_profile::StorageProfile;
 use crate::db::storage_profile_repository::StorageProfileRepository;
 use crate::events::rabbitmq::RabbitMqPublisher;
 use crate::helpers::data_error::DataError;
-
-#[async_trait]
-pub trait StorageProfileService: Send + Sync {
-    async fn save_storage_profile (&self, event: UserCreatedEvent) -> Result<StorageProfile, DataError>;
-    async fn toggle_storage_profile(&self, event: UserBlockedEvent) -> Result<(), DataError>;
-    async fn get_by_id(&self, id: Uuid) -> Result<Option<StorageProfile>, DataError>;
-    async fn reduce_taken_storage(&self, id: Uuid, size: i64) -> Result<(), DataError>;
-}
+use crate::service::contract::sp_service::StorageProfileService;
 
 #[derive(new)]
 pub struct StorageProfileServiceImpl {
@@ -38,11 +31,10 @@ impl StorageProfileService for StorageProfileServiceImpl {
     async fn toggle_storage_profile(&self, event: UserBlockedEvent) -> Result<(), DataError> {
         let profile = self.storage_profile_repo.get_by_id(event.user_id).await
             .map_err(|_| DataError::EntityNotFoundException("Storage Profile".to_string()))?;
-        
+
         let _ = self.storage_profile_repo.toggle_blocked(profile.unwrap(), event.is_deleted).await;
 
         Ok(())
-
     }
 
     async fn get_by_id(&self, id: Uuid) -> Result<Option<StorageProfile>, DataError> {
